HH-80537 rename package core -> starter
