+++ nab-example/src/main/java/ru/hh/nab/example/ExampleMain.java
@@ -2,7 +2,7 @@ package ru.hh.nab.example;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import ru.hh.nab.core.NabApplication;
+import ru.hh.nab.starter.NabApplication;
 
 @Configuration
 @Import(ExampleProdConfig.class)
