+++ nab-example/src/main/java/ru/hh/nab/example/ExampleProdConfig.java
@@ -3,7 +3,7 @@ package ru.hh.nab.example;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import ru.hh.nab.core.CoreProdConfig;
+import ru.hh.nab.starter.CoreProdConfig;
 
 @Configuration
 @Import({CoreProdConfig.class})
