+++ nab-starter-tests/pom.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>ru.hh.nab</groupId>
         <artifactId>nuts-and-bolts-parent</artifactId>
@@ -7,7 +8,6 @@
         <relativePath>..</relativePath>
     </parent>
 
-    <modelVersion>4.0.0</modelVersion>
     <packaging>jar</packaging>
     <artifactId>nab-starter-tests</artifactId>
