+++ nab-starter-tests/src/test/java/ru/hh/nab/starter/NabApplicationTest.java
@@ -1,0 +1,16 @@
+package ru.hh.nab.starter;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+import org.junit.Test;
+
+public class NabApplicationTest {
+
+  @Test
+  public void runShouldStartJetty() {
+    NabApplicationContext context = (NabApplicationContext) NabApplication.run(CoreTestConfig.class);
+
+    assertTrue(context.isServerRunning());
+    assertEquals(CoreTestConfig.TEST_SERVICE_NAME, context.getBean("serviceName"));
+  }
+}
