--- nab-starter/src/main/java/ru/hh/nab/core/NabApplicationContext.java
@@ -1,17 +1,0 @@
-package ru.hh.nab.core;
-
-import org.eclipse.jetty.server.Server;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-class NabApplicationContext extends AnnotationConfigApplicationContext {
-
-  private volatile Server server;
-
-  boolean isServerRunning() {
-    return server.isRunning();
-  }
-
-  void setServer(Server server) {
-    this.server = server;
-  }
-}
