--- nab-starter/src/main/java/ru/hh/nab/core/filters/RequestHeaders.java
@@ -1,6 +1,0 @@
-package ru.hh.nab.core.filters;
-
-public interface RequestHeaders {
-  String REQUEST_ID = "x-request-id";
-  String REQUEST_ID_DEFAULT = "NoRequestId";
-}
