--- nab-starter/src/main/java/ru/hh/nab/core/servlet/ServletConfig.java
@@ -1,16 +1,0 @@
-package ru.hh.nab.core.servlet;
-
-import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.glassfish.jersey.server.ResourceConfig;
-import org.springframework.context.ApplicationContext;
-
-public interface ServletConfig {
-
-  default String getServletMapping() {
-    return "/*";
-  }
-
-  void configureServletContext(ServletContextHandler servletContextHandler, ApplicationContext applicationContext);
-
-  ResourceConfig createResourceConfig(ApplicationContext context);
-}
