+++ nab-starter/src/main/java/ru/hh/nab/starter/StatsResource.java
@@ -1,0 +1,14 @@
+package ru.hh.nab.starter;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+
+@Path("/stats")
+public class StatsResource {
+  @GET
+  @Produces("text/csv")
+  public String stats() {
+    return "status,ok";
+  }
+}
