+++ nab-starter/src/main/java/ru/hh/nab/starter/filters/RequestHeaders.java
@@ -1,0 +1,6 @@
+package ru.hh.nab.starter.filters;
+
+public interface RequestHeaders {
+  String REQUEST_ID = "x-request-id";
+  String REQUEST_ID_DEFAULT = "NoRequestId";
+}
