+++ nab-testbase/pom.xml
@@ -1,4 +1,5 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>ru.hh.nab</groupId>
         <artifactId>nuts-and-bolts-parent</artifactId>
@@ -6,7 +7,6 @@
         <relativePath>..</relativePath>
     </parent>
 
-    <modelVersion>4.0.0</modelVersion>
     <artifactId>nab-testbase</artifactId>
     <packaging>jar</packaging>
